Abstract accelerometer to motion sensor
