+++ config/generic-mellow-super-infinty-hv.cfg
@@ -110,7 +110,8 @@ max_temp: 200
 #...
 
 # Accelerometer port, has <5V>,<GND>,<MISO>,<MOSI>,<INT>,<CS>,<SCLK>
-#[adxl345]
+#[motion_sensor]
+#chip: adxl345
 #spi_speed: 5000
 #spi_software_miso_pin: PD0
 #spi_software_mosi_pin: PD1
