+++ config/sample-bigtreetech-hermit-crab-canbus.cfg
@@ -9,7 +9,8 @@
 serial: /dev/serial/by-id/usb-Klipper_Klipper_firmware_12345-if00
 #canbus_uuid: 0e0d81e4210c
 
-[adxl345]
+[motion_sensor]
+chip: adxl345
 cs_pin: HermitCrab: PB12
 spi_bus: spi2
 axes_map: y,z,-x
