+++ test/klippy/input_shaper.cfg
@@ -73,10 +73,11 @@ shaper_freq_x: 33.2
 shaper_type_x: ei
 shaper_freq_x: 39.3
 
-[adxl345]
+[motion_sensor]
+chip: adxl345
 cs_pin: PK7
 axes_map: -x,-y,z
 
 [resonance_tester]
 probe_points: 20,20,20
-accel_chip: adxl345
+accel_chip: motion_sensor
